Refactor commands and output
