+++ src/commands/init.rs
@@ -1,0 +1,42 @@
+use crate::{
+    commands::DotCommand,
+    error::DotError,
+    fs::{symlink::SymLinkOperations, FileSystem},
+    manifest::ManifestOperations,
+    output::CommandOutput,
+    service::DotService,
+};
+
+/// Command to initialize a new dot repository
+pub struct InitCommand<F: FileSystem, S: SymLinkOperations, M: ManifestOperations, O: CommandOutput>
+{
+    service: DotService<F, S, M>,
+    output: O,
+}
+
+impl<
+        F: FileSystem + Clone,
+        S: SymLinkOperations + Clone,
+        M: ManifestOperations + Clone,
+        O: CommandOutput + Clone,
+    > InitCommand<F, S, M, O>
+{
+    pub fn new(service: DotService<F, S, M>, output: O) -> Self {
+        Self { service, output }
+    }
+}
+
+impl<
+        F: FileSystem + Clone,
+        S: SymLinkOperations + Clone,
+        M: ManifestOperations + Clone,
+        O: CommandOutput + Clone,
+    > DotCommand for InitCommand<F, S, M, O>
+{
+    fn execute(&self) -> Result<(), DotError> {
+        self.service.init()?;
+        self.output
+            .display_success("Initialized empty dot repository");
+        Ok(())
+    }
+}
