+++ src/main.rs
@@ -3,6 +3,7 @@ mod commands;
 mod error;
 mod fs;
 mod manifest;
+mod output;
 mod path_ext;
 mod service;
